+++ src/main/java/keysson/apis/validacao/mapper/UserRowMapper.java
@@ -14,8 +14,8 @@ public class UserRowMapper implements RowMapper<User> {
     @Override
     public User mapRow(ResultSet rs, int rowNum) throws SQLException {
         User user = new User();
-        user.setId(rs.getLong("id"));
-        user.setCompanyId(rs.getLong("company_id"));
+        user.setId(rs.getInt("id"));
+        user.setCompanyId(rs.getInt("company_id"));
         user.setUsername(rs.getString("username"));
         user.setPassword(rs.getString("password"));
         user.setStatus(rs.getInt("status"));
